refactor : 테스트 정리 및 포트 충돌 해결
