+++ src/test/java/roomescape/repository/JdbcReservationTimeRepositoryTest.java
@@ -8,20 +8,23 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.time.entity.ReservationTime;
 import roomescape.time.repository.JdbcReservationTimeRepository;
 
-@SpringBootTest
-@Transactional
+@JdbcTest
 public class JdbcReservationTimeRepositoryTest {
 
-    @Autowired
     private JdbcReservationTimeRepository jdbcReservationTimeRepository;
 
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
     @BeforeEach
     void setup() {
+        jdbcReservationTimeRepository = new JdbcReservationTimeRepository(jdbcTemplate);
         ReservationTime nonIdReservationTime = ReservationTime.createNew(LocalTime.parse("10:00"));
         jdbcReservationTimeRepository.save(nonIdReservationTime);
     }
