+++ src/test/java/roomescape/validate/ReservationTimeRequestValidationTest.java
@@ -1,17 +1,28 @@
-package roomescape.api.validate;
+package roomescape.validate;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.HashMap;
 import java.util.Map;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.annotation.DirtiesContext;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class ReservationTimeRequestValidationTest {
+
+    @LocalServerPort
+    int port;
+
+    @BeforeEach
+    void setup(){
+        RestAssured.port = port;
+    }
 
     @Test
     @DisplayName("시간 비어있으면 예약 시간 생성 실패 (400)")
